Remove a unnecessary else in DefaultSingletonBeanRegistry
The method `getSingleton(String, ObjectFactory<?>)` may throw a
`BeanCurrentlyInCreationException` in multiple threads.

If Thread-A acquires the `singletonLock`, but before it assigns itself to
`singletonCreationThread`,
Thread-B gets a null value at `singletonCreationThread`,
then Thread-A and Thread-B will both invoke `beforeSingletonCreation`.

In this commit, I remove the else,
so the thread which does not acquire `singletonLock` will be blocked
at the code `this.singletonLock.lock()`.

Closes gh-33463