Remove an unnecessary else in DefaultSingletonBeanRegistry
The method `getSingleton(String, ObjectFactory<?>)` may throw a
`BeanCurrentlyInCreationException` in multiple threads.

If Thread-A acquires the `singletonLock`, and it assigns itself to
`singletonCreationThread` immediately,
Thread-B will get a not null value at `singletonCreationThread`,
then Thread-A and Thread-B will both invoke `beforeSingletonCreation`.

In this commit, I remove the else,
so the thread which does not acquire `singletonLock` will be blocked
at the code `this.singletonLock.lock()`.

Closes gh-33463