Add heartbeat (ping) functionality to SseEmitter
This commit introduces the heartbeat (ping) feature in SseEmitter. It allows the server to send periodic heartbeat messages to the client at a specified interval to keep the connection alive and detect client disconnects. Heartbeat messages are sent as comments (":heartbeat") and stop when the emitter is completed, times out, or encounters an error.

See: gh-33355