Refactor -> StringUtils replace Optimization
The += 16 used before is too abstract.

-If NewPattern is longer than 16, StringBuilder must be re -permitted again.
-If NewPattern is shorter than 16, you will be assigned a memory to StringBuilder.

For this reason, we will request a Code that is optimized.

There is also a possibility that the length of the + NewPattern will be abnormally longer,
To prevent the maximum size assignment, the maximum value is set to the maximum value of the int range.