Revisit GitHub issue template
Prior to this commit, this repository had a single issue template
markdown file that listed all the information for creating a new issue
in this project. Because there was a lot, people missed warnings against
creating security reports publicly or asking questions in the issue
tracker.

This commit introduces an issue template configuration that provides
developers with a choice before creating the actual issue.

Closes gh-33711